# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PeakNovel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
  /users/{id}/follow:
    post:
      operationId: followUser
      tags: [users]
      summary: Follow an author
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Followed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowStatus"
    delete:
      operationId: unfollowUser
      tags: [users]
      summary: Unfollow an author
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unfollowed

  # ── PREMIUM ────────────────────────────────────────────
...
components:
  schemas:
...
    AuthUser:
      type: object
      required: [id, username, email, role]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        isPremium:
          type: boolean
        followerCount:
          type: integer
        followingCount:
          type: integer
        bookCount:
          type: integer
        createdAt:
          type: string

    # ── HOME SCHEMAS ──────────────────────────────────────
...
    Book:
      type: object
      required: [id, title, authorId, authorName, genre, status, coverUrl, viewCount, chapterCount, rating, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatar:
          type: ["string", "null"]
        genre:
          type: string
        tags:
          type: array
          items:
            type: string
        synopsis:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        status:
          type: string
        viewCount:
          type: integer
        chapterCount:
          type: integer
        rating:
          type: number
        ratingCount:
          type: integer
        isPremium:
          type: boolean
        isFeatured:
          type: boolean
        isBookmarked:
          type: ["boolean", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    UserProfile:
      type: object
      required: [id, username, role, followerCount, followingCount, bookCount, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        role:
          type: string
        isPremium:
          type: boolean
        followerCount:
          type: integer
        followingCount:
          type: integer
        bookCount:
          type: integer
        books:
          type: array
          items:
            $ref: "#/components/schemas/Book"
        isFollowing:
          type: ["boolean", "null"]
        createdAt:
          type: string
    ProfileUpdate:
      type: object
      properties:
        username:
          type: string
        bio:
          type: string
        avatar:
          type: string
...
    FollowStatus:
      type: object
      required: [following]
      properties:
        following:
          type: boolean
...